move functions display_list_tasks and display_list_models
