+++ scripts/sct_deepseg.py
@@ -18,7 +18,6 @@ import argparse
 import os
 import sys
 
-import colored
 import spinalcordtoolbox as sct
 import spinalcordtoolbox.deepseg.core
 import spinalcordtoolbox.deepseg.models
@@ -116,56 +115,6 @@ def get_parser():
     return parser
 
 
-def display_list_models():
-    models = sct.deepseg.models.list_models()
-    # Display beautiful output
-    color = {True: 'green', False: 'red'}
-    default = {True: '[*]', False: ''}
-    print("{:<25s}DESCRIPTION".format("MODEL"))
-    print("-" * 80)
-    for name_model, value in models.items():
-        path_model = sct.deepseg.models.folder(name_model)
-        print("{}{}".format(
-            colored.stylize(''.join((name_model, default[value['default']])).ljust(25),
-                            colored.fg(color[sct.deepseg.models.is_valid(path_model)])),
-            colored.stylize(value['description'],
-                            colored.fg(color[sct.deepseg.models.is_valid(path_model)]))
-        ))
-    print(
-        '\nLegend: {} | {} | default: {}\n'.format(
-            colored.stylize("installed", colored.fg(color[True])),
-            colored.stylize("not installed", colored.fg(color[False])),
-            default[True]))
-    exit(0)
-
-
-def display_list_tasks():
-    tasks = sct.deepseg.models.list_tasks()
-    # Display beautiful output
-    color = {True: 'green', False: 'red'}
-    default = {True: '[*]', False: ''}
-    print("{:<20s}{:<50s}MODELS".format("TASK", "DESCRIPTION"))
-    print("-" * 80)
-    for name_task, value in tasks.items():
-        path_models = [sct.deepseg.models.folder(name_model) for name_model in value['models']]
-        are_models_valid = [sct.deepseg.models.is_valid(path_model) for path_model in path_models]
-        task_status = colored.stylize(name_task.ljust(20),
-                                      colored.fg(color[all(are_models_valid)]))
-        description_status = colored.stylize(value['description'].ljust(50),
-                                             colored.fg(color[all(are_models_valid)]))
-        models_status = ', '.join([colored.stylize(model_name,
-                                                   colored.fg(color[is_valid]))
-                                   for model_name, is_valid in zip(value['models'], are_models_valid)])
-        print("{}{}{}".format(task_status, description_status, models_status))
-
-    print(
-        '\nLegend: {} | {} | default: {}\n'.format(
-            colored.stylize("installed", colored.fg(color[True])),
-            colored.stylize("not installed", colored.fg(color[False])),
-            default[True]))
-    exit(0)
-
-
 def main():
     parser = get_parser()
     args = parser.parse_args(args=None if sys.argv[1:] else ['--help'])
@@ -173,11 +122,11 @@ def main():
 
     # Deal with model
     if args['list_models']:
-        display_list_models()
+        sct.deepseg.models.display_list_models()
 
     # Deal with task
     if args['list_tasks']:
-        display_list_tasks()
+        sct.deepseg.models.display_list_tasks()
 
     if 'install_model' in args:
         sct.deepseg.models.install_model(args['install_model'])
