+++ spinalcordtoolbox/deepseg/models.py
@@ -7,6 +7,7 @@ Deals with models for deepseg module. Available models are listed under MODELS.
 import os
 import json
 import logging
+import colored
 
 import spinalcordtoolbox as sct
 import spinalcordtoolbox.download
@@ -108,12 +109,62 @@ def list_models():
     return {name: value for name, value in MODELS.items()}
 
 
+def display_list_models():
+    models = sct.deepseg.models.list_models()
+    # Display beautiful output
+    color = {True: 'green', False: 'red'}
+    default = {True: '[*]', False: ''}
+    print("{:<25s}DESCRIPTION".format("MODEL"))
+    print("-" * 80)
+    for name_model, value in models.items():
+        path_model = sct.deepseg.models.folder(name_model)
+        print("{}{}".format(
+            colored.stylize(''.join((name_model, default[value['default']])).ljust(25),
+                            colored.fg(color[sct.deepseg.models.is_valid(path_model)])),
+            colored.stylize(value['description'],
+                            colored.fg(color[sct.deepseg.models.is_valid(path_model)]))
+        ))
+    print(
+        '\nLegend: {} | {} | default: {}\n'.format(
+            colored.stylize("installed", colored.fg(color[True])),
+            colored.stylize("not installed", colored.fg(color[False])),
+            default[True]))
+    exit(0)
+
+
 def list_tasks():
     """
     Display available tasks with description.
     :return: dict: Tasks that are installed
     """
     return {name: value for name, value in TASKS.items()}
+
+
+def display_list_tasks():
+    tasks = sct.deepseg.models.list_tasks()
+    # Display beautiful output
+    color = {True: 'green', False: 'red'}
+    default = {True: '[*]', False: ''}
+    print("{:<20s}{:<50s}MODELS".format("TASK", "DESCRIPTION"))
+    print("-" * 80)
+    for name_task, value in tasks.items():
+        path_models = [sct.deepseg.models.folder(name_model) for name_model in value['models']]
+        are_models_valid = [sct.deepseg.models.is_valid(path_model) for path_model in path_models]
+        task_status = colored.stylize(name_task.ljust(20),
+                                      colored.fg(color[all(are_models_valid)]))
+        description_status = colored.stylize(value['description'].ljust(50),
+                                             colored.fg(color[all(are_models_valid)]))
+        models_status = ', '.join([colored.stylize(model_name,
+                                                   colored.fg(color[is_valid]))
+                                   for model_name, is_valid in zip(value['models'], are_models_valid)])
+        print("{}{}{}".format(task_status, description_status, models_status))
+
+    print(
+        '\nLegend: {} | {} | default: {}\n'.format(
+            colored.stylize("installed", colored.fg(color[True])),
+            colored.stylize("not installed", colored.fg(color[False])),
+            default[True]))
+    exit(0)
 
 
 def get_metadata(folder_model):
